refactor: merge recipes controllers into a single file and change the recipesServices.js
